requirements.txt
requests
beautifulsoup4
python-dotenv
schedule
loguru
playwright
langchain
langchain-community
groq 